requirements.txt
# SehatScan Core Dependencies
streamlit>=1.39.0
plotly>=5.24.0
huggingface-hub>=0.24.0
python-dotenv>=1.0.1
pillow>=10.0.0
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
google-generativeai>=0.8.0

# OCR Dependencies (optional - app works without OCR using JSON input)
paddlepaddle-cpu>=2.6.0
paddleocr>=2.7.0
opencv-python-headless>=4.8.0